+++ react/AddressRules.js
@@ -12,6 +12,8 @@ class AddressRules extends Component {
     this.state = {
       country: null,
       rules: null,
+      loadingRules: false,
+      error: null,
     }
   }
 
@@ -35,14 +37,19 @@ class AddressRules extends Component {
   }
 
   fetchRules(rulePromise) {
+    this.setState({ loadingRules: true })
     return rulePromise
-      .then(ruleData => ruleData.default || ruleData)
-      .catch(error => {
-        const errorType = this.parseError(error)
-        if (errorType) {
+      .then((ruleData) => {
+        this.setState({ error: null })
+        return ruleData.default || ruleData
+      })
+      .catch((error) => {
+        const notFoundErrorType = this.parseError(error)
+
+        if (notFoundErrorType) {
           if (process.env.NODE_ENV !== 'production') {
             console.warn(
-              `Couldn't load rules for country ${errorType}, using default rules instead.`,
+              `Couldn't load rules for country ${notFoundErrorType}, using default rules instead.`,
             )
           }
           return defaultRules
@@ -51,6 +58,17 @@ class AddressRules extends Component {
         if (process.env.NODE_ENV !== 'production') {
           console.error('An unknown error occurred.', error)
         }
+
+        if (this.props.useDefaultRulesAsFallback) {
+          // Since not found rules can happen quite frequently,
+          // only unexpected errors should notify the consumers.
+          this.setState({ error })
+
+          return defaultRules
+        }
+      })
+      .finally(() => {
+        this.setState({ loadingRules: false })
       })
   }
 
@@ -69,7 +87,7 @@ class AddressRules extends Component {
         // set a hidden flag for internal usage
         _usingGeolocationRules: true,
         // overwrite field with configs defined on `rules.geolocation`
-        fields: rules.fields.map(field => {
+        fields: rules.fields.map((field) => {
           if (rules.geolocation[field.name]) {
             // ignore unrelated props for the field
             // eslint-disable-next-line no-unused-vars
@@ -89,12 +107,21 @@ class AddressRules extends Component {
 
   render() {
     const { children } = this.props
-    const { rules } = this.state
+    const { rules, loadingRules, error } = this.state
 
     if (!rules) return null
 
     return (
-      <RulesContext.Provider value={rules}>{children}</RulesContext.Provider>
+      <RulesContext.Provider
+        value={{
+          rules,
+          loadingRules,
+          fetchRules: this.updateRules.bind(this),
+          rulesError: error,
+        }}
+      >
+        {children}
+      </RulesContext.Provider>
     )
   }
 }
@@ -107,6 +134,8 @@ AddressRules.propTypes = {
   shouldUseIOFetching: PropTypes.bool,
   /** Whether the rules should contemplate the geolocation field rules */
   useGeolocation: PropTypes.bool,
+  /** Whether to always use the default rules as fallback or not */
+  useDefaultRulesAsFallback: PropTypes.bool,
 }
 
 export default AddressRules